+++ server/src/main/java/com/inventage/portal/gateway/core/entrypoint/Entrypoint.java
@@ -107,7 +107,7 @@ public class Entrypoint {
         enabled = false;
     }
 
-    class Tls {
+    static class Tls {
         public JksOptions jksOptions() {
             return null;
         }
@@ -144,7 +144,7 @@ public class Entrypoint {
                             .handler((Handler<RoutingContext>) mf.result()));
             LOGGER.info("EntryMiddlewares created successfully");
         }).onFailure(cfErr -> {
-            final String errMsg = String.format("Failed to create EntryMiddlewares");
+            final String errMsg = "Failed to create EntryMiddlewares";
             LOGGER.warn("{}", errMsg);
         });
     }