+++ server/src/main/java/com/inventage/portal/gateway/proxy/config/label/Parser.java
@@ -51,7 +51,6 @@ public class Parser {
             for (String filter : filters) {
                 if (key.startsWith(filter)) {
                     filteredKeys.add(key);
-                    continue;
                 }
             }
         }
@@ -64,7 +63,7 @@ public class Parser {
             return decodedConf;
         }
         catch (IllegalArgumentException e) {
-            LOGGER.warn("Decoding to JSON failed: '{}'", e.getMessage(),
+            LOGGER.warn("Decoding to JSON failed: '{}' '{}'", e.getMessage(),
                 labels.toString());
         }
         return null;