+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/checkRoute/CheckRouteMiddlewareFactory.java
@@ -1,0 +1,32 @@
+package com.inventage.portal.gateway.proxy.middleware.checkRoute;
+
+import com.inventage.portal.gateway.proxy.config.dynamic.DynamicConfiguration;
+import com.inventage.portal.gateway.proxy.middleware.Middleware;
+import com.inventage.portal.gateway.proxy.middleware.MiddlewareFactory;
+import io.vertx.core.Future;
+import io.vertx.core.Vertx;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.Router;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * Factory for checkRoute Middleware.
+ */
+public class CheckRouteMiddlewareFactory implements MiddlewareFactory {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(CheckRouteMiddlewareFactory.class);
+
+    @Override
+    public String provides() {
+        return DynamicConfiguration.MIDDLEWARE_CHECK_ROUTE;
+    }
+
+    @Override
+    public Future<Middleware> create(Vertx vertx, Router router, JsonObject middlewareConfig) {
+        LOGGER.debug("Created '{}' middleware successfully",
+                DynamicConfiguration.MIDDLEWARE_CHECK_ROUTE);
+        return Future.succeededFuture(new CheckRouteMiddleware());
+    }
+
+}