+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/proxy/ProxyMiddleware.java
@@ -31,10 +31,10 @@ public class ProxyMiddleware implements Middleware {
     private static final String X_FORWARDED_HOST = "X-Forwarded-Host";
     private static final String X_FORWARDED_PORT = "X-Forwarded-Port";
 
-    private HttpProxy httpProxy;
+    private final HttpProxy httpProxy;
 
-    private String serverHost;
-    private int serverPort;
+    private final String serverHost;
+    private final int serverPort;
 
     private List<Handler<StringBuilder>> incomingRequestURIModifiers;
     private List<Handler<MultiMap>> outgoingResponseHeadersModifiers;